Feature filter component - material UI:
 - filter component.
