+++ client/scripts/Components/drawer/drawer.component.js
@@ -1,8 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import throttle from 'lodash.throttle';
+import axios from 'axios';
+import moment from 'moment';
 
 import AppBarComponent from '../app-bar/app-bar.component';
+import FilterComponent from '../filter/filter.component';
+
+import { ENDPOINT } from '../../../../config/constants';
 
 // begin of material UI //
 import { withStyles } from '@material-ui/core/styles';
@@ -13,8 +19,6 @@ import Divider from '@material-ui/core/Divider';
 import IconButton from '@material-ui/core/IconButton';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import ChevronRightIcon from '@material-ui/icons/ChevronRight';
-import FileProvider from '../../Providers/FileProvider';
-import FileContainer from '../../Containers/FileContainer';
 // end of material UI //
 
 const drawerWidth = 345;
@@ -54,13 +58,204 @@ const styles = (theme) => ({
     marginLeft: 0,
   },
   grow: {
-    width: '100%'
-  }
+    width: '100%',
+  },
 });
+
+const INITIAL_STATE = {
+  files: [],
+  deletedSize: 0,
+  paging: {
+    total: 0,
+    page: 1,
+    pages: 1,
+  },
+  currentPage: 1,
+  rate_time: 0,
+  rate_count: 0,
+  hasRun: false,
+  hasFiles: false,
+  searchDetails: {
+    from: null,
+    to: null,
+    types: null,
+    channel: null,
+  },
+};
 
 class PersistentDrawerLeft extends React.Component {
   state = {
     open: false,
+    ...INITIAL_STATE,
+  };
+
+  componentWillUnmount() {
+    clearInterval(this.myTimer);
+  }
+
+  getFiles = (from = null, to = null, types = null, channel = null) => {
+    const now = moment()
+      .seconds(0)
+      .milliseconds(0)
+      .minutes(0);
+    let fromValue, toValue;
+    if (from) {
+      fromValue = from ? moment(from).unix() : null;
+    }
+
+    if (to) {
+      toValue = !moment(now).isSame(to) ? moment(to).unix() : null;
+    }
+
+    this.setState(
+      {
+        searchDetails: {
+          from: fromValue,
+          to: toValue,
+          types,
+          channel,
+        },
+      },
+      () => {
+        this.callGetFiles();
+      },
+    );
+  };
+
+  startTimer = () => {
+    if (this.state.rate_count >= 25 && this.state.rate_time) {
+      this.props.updateError(
+        'Slow down that trigger finger, Slack has a rate limit of 50 calls every minute.',
+      );
+      return false;
+    }
+
+    if (!this.state.rate_time) {
+      this.myTimer = setTimeout(this.timer.bind(this), 15000);
+      this.setState({ rate_time: true, rate_count: 1 });
+    }
+
+    return true;
+  };
+
+  timer = () => {
+    this.setState({ rate_time: false, rate_count: 0 });
+  };
+
+  // ### TODO Refactor the shit out of this
+  callGetFiles = throttle(async () => {
+    if (await !this.startTimer()) {
+      return;
+    }
+
+    try {
+      const res = await axios.get(`${ENDPOINT}files.list`, {
+        params: {
+          token: this.props.accessToken,
+          user: !this.props.isAdmin ? this.props.userId : null,
+          ts_from: this.state.searchDetails.from,
+          ts_to: this.state.searchDetails.to,
+          page: this.state.currentPage,
+          types: this.state.searchDetails.types.length
+            ? this.state.searchDetails.types
+            : null,
+          channel: this.state.searchDetails.channel.length
+            ? this.state.searchDetails.channel
+            : null,
+        },
+      });
+
+      if (!res.data.ok) {
+        this.props.updateError('Slack says no :(', 'GetFiles returned not ok');
+        return;
+      }
+
+      // If results are completely empty, reset the page count
+      if (!res.data.files.length) {
+        this.setState({
+          files: [],
+          hasFiles: false,
+          hasRun: true,
+          currentPage: 1,
+          rate_count: this.state.rate_count + 1,
+        });
+        return;
+      }
+
+      this.setState({
+        files: res.data.files,
+        hasFiles: res.data.files.length > 0,
+        hasRun: true,
+        paging: res.data.paging,
+        rate_count: this.state.rate_count + 1,
+      });
+    } catch (err) {
+      this.props.updateError(
+        'Slack looks like it is down :(',
+        `getFiles - ${err}`,
+      );
+      this.setState({
+        files: [],
+        paging: INITIAL_STATE.paging,
+        rate_count: this.state.rate_count + 1,
+      });
+    }
+  }, 100);
+
+  handlePageUpdate = (pageNumber) => {
+    this.setState({ currentPage: pageNumber }, () => {
+      this.callGetFiles();
+    });
+  };
+
+  callDeleteFile = throttle(async (fileId) => {
+    if (!this.startTimer()) {
+      return;
+    }
+
+    try {
+      const res = await axios.get(`${ENDPOINT}files.delete`, {
+        params: {
+          token: this.props.accessToken,
+          file: fileId,
+        },
+      });
+      this.setState({ rate_count: this.state.rate_count + 1 });
+
+      if (!res.data.ok) {
+        this.props.updateError(
+          'Slack said no :(. Try it again',
+          'callDeleteFile - res was not ok',
+        );
+      } else {
+        this.deleteFile(fileId);
+      }
+    } catch (err) {
+      this.props.updateError(
+        'You must be logged in!',
+        `callDeleteFile - ${err}`,
+      );
+    }
+  }, 1000);
+
+  deleteFile = (fileId) => {
+    const file = this.state.files.filter((item) => item.id === fileId)[0];
+    const filteredFiles = this.state.files.filter((item) => item.id !== fileId);
+    const fileSize = file.size + this.state.deletedSize;
+    if (!filteredFiles.length) {
+      this.setState({
+        files: [],
+        error: INITIAL_STATE.error,
+        deletedSize: fileSize,
+      });
+      return;
+    }
+
+    this.setState({
+      files: filteredFiles,
+      error: INITIAL_STATE.error,
+      deletedSize: fileSize,
+    });
   };
 
   handleDrawerOpen = () => {
@@ -87,7 +282,8 @@ class PersistentDrawerLeft extends React.Component {
       accessToken,
       channels,
       userId,
-      updateError = () => {},
+      updateError = () => {
+      },
     } = this.props;
 
     const { open } = this.state;
@@ -123,18 +319,12 @@ class PersistentDrawerLeft extends React.Component {
           </div>
           <Divider/>
 
-          {/*begin of filter component*/}
-          <FileProvider
+          <FilterComponent
             isLoggedIn={isLoggedIn}
-            userId={userId}
-            accessToken={accessToken}
-            isAdmin={isAdmin}
             channels={channels}
+            onGetFiles={this.getFiles}
             updateError={updateError}
-          >
-            <FileContainer />
-          </FileProvider>
-          {/* end of filter component*/}
+          />
 
         </Drawer>
         <main
@@ -143,6 +333,8 @@ class PersistentDrawerLeft extends React.Component {
           })}
         >
           <div className={classes.drawerHeader}/>
+
+
           <Typography paragraph>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
             incididunt ut labore et dolore magna aliqua. Rhoncus dolor purus non enim praesent
