+++ client/scripts/Containers/FileContainer.js
@@ -1,61 +1,9 @@
 import React, { Component, Fragment } from 'react';
 
-import { formatBytes } from '../utils';
 import { FileContext } from '../Providers/FileProvider';
 import Form from './Form';
-import ButtonComponent from '../Components/button/button.component';
-import FAQ from '../Components/FAQ';
 
 class FileContainer extends Component {
-  state = {
-    showFAQ: false,
-  };
-
-  toggleFAQ = () => {
-    this.setState({ showFAQ: !this.state.showFAQ });
-  };
-
-  showFAQ = () => {
-    if (!this.state.showFAQ) {
-      return null;
-    }
-
-    return (
-      <div className="FAQ">
-        <div className="FAQ__Wrapper">
-          <FAQ onClose={this.toggleFAQ} />
-        </div>
-      </div>
-    );
-  };
-
-  displayBar = (deletedSize, hasRun) => {
-    return (
-      <div className="FileWrapper__Details">
-        <div>
-          {deletedSize > 0 && hasRun ? (
-            <p className="Count__Text">
-              Nice! You just saved{' '}
-              <span className="red">{formatBytes(deletedSize)}</span>
-            </p>
-          ) : null}
-        </div>
-        <div className="FileWrapper__Details-share">
-          <p>
-            <a
-              href="https://twitter.com/intent/tweet?text=Delete%20and%20manage%20files%20from%20your%20Slack%20workspace%20with%20Slack%20Cleaner&url=https://www.slackcleaner.com&hashtags=slack&via=drewisthe"
-              className="twitter-share-button Button"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Tweet about Slack Cleaner
-            </a>
-          </p>
-          <ButtonComponent title="Questions? FAQ" icon={'Help'} color={'default'} onClick={this.toggleFAQ} />
-        </div>
-      </div>
-    );
-  };
 
   render() {
     return (
@@ -64,7 +12,7 @@ class FileContainer extends Component {
           <Fragment>
             <main className="MainContent cf">
               <Form
-                getFiles={context.getFiles}
+                getFiles={context.files}
                 channels={context.channels}
                 isLoggedIn={context.isLoggedIn}
               />
